Move to a pipe delimited list for rule lists:
A pipe delimited doesn't carry with it the requirement
that all JSON rules be devoid of space. Potential
improved user experience. Also, update the doc.
